refactor: move stories nav to component
